Revisited licenses for included fonts. Arial font was grabbed from my Windows 10 computer, but it turns out, it definitely did NOT carry a public license that let me redistribute it. When I included the other fonts in the assets folder, I only included the font file. There was no license information for each included font.
As such, I started over. I deleted all of the font files in the assets folder. I downloaded fonts from 1001fonts.com that were specifically
licensed as "free for commercial" and I verified language in their included documentation that stated the same thing. I included, not only
the font.ttf files, but the documentation for each font that states that I have the right to redistribute their fonts for commercial purposes.

Further, I changed the default font from Arial to Questrial-Regular, a sans-serif font similar to Arial. This change was made in the
plugin_constants.php file. Additionally, I changed the Options page to display the 4 available fonts, now. And tested the creation of
the screenshot.png file with the new fonts.

All is working.