Edit file capabilities completed. Never discovered why the display window is tiny on some computers. Worked around it: in Javascript, during the load of the file being edited, I reset the size of the id='wpbody-content' div (WP core) to 70% of the screen height.
Working on making the edit file capabilities responsive.

Will only edit .php, .css, .js files now. Need to add the ability to add additional file types as an option.

Added hexdump functionality (javascript) when in WP_DEBUG=true.

Need to add file name at the top of the screen and fix spacing issues with regard to the file placement on the screen.