Intermediate commit.
Previous commit was for a menu page. Those changes have been removed.

The primary dashboard page for the Child Themes Helper has changed to a tab driven page.
Tabs will remain at the top of the page, while the content scrolls beneath the tabs.

Copying files from parent to child or removing files from the child theme will exist on a separate tab from the ability to edit files.
I didn't like the clunky nature of using a left click to copy and a right click to edit.