Version updated to 2.2.5
When testing for WordPress 6.0, I caused a bug which crashed WordPress.
The plugin prevents users from deleting a child theme's styles.css file. But a user could modify styles.css and delete the header. That's what I did. That causes the WordPress site to crash.

I added code which causes the "active theme" as defined by the Child Themes Helper plugin, to be removed in the case of a Child Theme which lacks the appropriate header information.

Found a problem where certain files couldn't be copied. This is due to a problem in the WordPress sanitize_file_name() function. When a file name has multiple 'extensions', as in This-is-my-file.ext.ext, the sanitize function appends an underscore to the end of the first (or subsequent, but not the last) extension. This causes the file to not copy.

I modified my version of the santize function to remove this feature.

Added code to write a message to the error log the next time a source file or target folder do not exist.

I used to edit these files with EditPlus, an FTP-enabled file editor. When a new release of their software removed the capability of the user to control whether it write out linux line endings or windows line endings, I started using Microsoft Visual Studio Code for all editing. Unfortunately, VSCode sees all of the previously uneditted files as having a blank line between lines of code. I am still finding and removing these when I find them.

I have tested the Child Themes Helper with WordPress 6.0 on php 7.4.1. Now I will test with php 8.x, and hopefully not find anything else previously overlooked problems.